refactor: explore data using pipeline instance
